@@ -1,15 +1,7 @@
 import './app.css'
 import App from './App.svelte'
+import { mount } from 'svelte'
 
-let app: App | null = null;
-const targetElement = document.getElementById('app');
+const app = mount(App, { target: document.getElementById("app") as HTMLElement })
 
-if (targetElement) {
-  app = new App({
-    target: targetElement,
-  });
-} else {
-  console.error('Target element not found');
-}
-
-export default app;
+export default app
